Bugfix parallelisation: missing reduce of XMCT (strata run time)
In 4fa782a8ba6d01bb274478b242d9d9a67e9aad8c a bug was introduced that stops the reduce of the XMCT variable.
A result of this was a wrong distribution of MPI processes on each strata for stratum No. > 1 in a second iteration, if using the proportional allocation schema for the parallelisation.