CMake: add git version hash output
Get git version hash information via cmake during build procress and
print the information at the beginning of the run to the standard
output unit.

In case cmake is not used, a dummy routine is provided with no output.
No changes to the Makefile are required except adding the file
src/assistant/version.f to the compile process.